requirements.txt
django-cors-headers==3.7.0
django_celery_beat==2.2.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
django-simple-captcha==0.5.13
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-extensions==0.7.0
jsonpath==0.82
mongoengine==0.22.1
Pillow==8.4.0
PyJWT==1.7.1
pymongo==3.11.3
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
coreapi==2.3.3
user-agents==2.2.0
eventlet==0.32.0
psutil==5.8.0
drf-yasg ==1.20.0
aioredis==1.1.0
kombu==5.1.0
amqp==5.1.1
asgiref==3.4.1
asn1crypto==0.24.0
autobahn==19.11.2
bcrypt==3.1.4
billiard==3.6.4.0
celery==5.1.2
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
daphne==2.3
diff-match-patch==20121119
Django==2.2.28
django-celery-results==2.1.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-reversion==2.0.13
django-timezone-field==4.1.2
et-xmlfile==1.0.1
future==0.15.2
gevent==20.4.0
gitdb2==2.0.5
GitPython==2.1.11
html5lib==1.0.1
httplib2==0.9.2
hyperlink==18.0.0
incremental==17.5.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
msgpack==0.6.0
msgpack-python==0.5.6
nose==1.3.7
odfpy==1.3.6
openpyxl==2.5.3
ply==3.11
pycryptodomex==3.7.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pysmi==0.3.0
pysnmp==4.4.4
python-crontab==2.3.5
python-dateutil==2.7.3
python-nmap==0.6.1
smmap2==2.0.5
svn==0.3.46
tablib==0.12.1
txaio==23.1.1
unicodecsv==0.14.1
vine==5.0.0
webencodings==0.5.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==6.0
xlsxwriter==1.1.8
cx-Oracle==7.3.0
django-pure-pagination==0.3.0
twisted==20.3.0
Automat==0.8.0
django-ninja==0.18.0
pydantic==1.9.1
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
elasticsearch7==7.17.0
email-validator==1.1.3
fastapi==0.67.0
greenlet==1.1.0
h11==0.12.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.31
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.7.2
frozenlist==1.2.0
click==7.1.2
multidict==5.2.0
django-jsonfield-backport==1.0.5
lxml==4.9.1
PyYAML==6.0
websockets==11.0.1
gunicorn==20.1.0
uvicorn[standard]==0.17.0
markupsafe==2.0.1
